# Controllers/CNCIOController.cs
using Microsoft.AspNetCore.Mvc;
using System.Collections.Generic;
using System.Threading.Tasks;
using HavenCNCServer.Models;

namespace HavenCNCServer.Controllers
{
    /// <summary>
    /// CNC IO Control - Handles inputs, outputs, vacuum, dust collection, and specialized IO
    /// </summary>
    [ApiController]
    [Route("api/[controller]")]
    public class CNCIOController : ControllerBase
    {
        #region Basic IO Control

        /// <summary>
        /// Get current input states
        /// </summary>
        /// <returns>Dictionary of input numbers and their states</returns>
        [HttpGet("GetCurrentInputs")]
        public async Task<IActionResult> GetCurrentInputs()
        {
            // TODO: Implement get current inputs
            await Task.Delay(1);
            var inputs = new Dictionary<int, bool>
            {
                { 1, true },
                { 2, false },
                { 3, true }
            };
            return Ok(new { inputs });
        }

        /// <summary>
        /// Get current output states
        /// </summary>
        /// <returns>Dictionary of output numbers and their states</returns>
        [HttpGet("GetCurrentOutputs")]
        public async Task<IActionResult> GetCurrentOutputs()
        {
            // TODO: Implement get current outputs
            await Task.Delay(1);
            var outputs = new Dictionary<int, bool>
            {
                { 1, false },
                { 2, true },
                { 3, false }
            };
            return Ok(new { outputs });
        }

        /// <summary>
        /// Check if specific input is active
        /// </summary>
        /// <param name="inputNumber">Input number to check</param>
        /// <returns>Input active status</returns>
        [HttpGet("IsInputActive/{inputNumber}")]
        public async Task<IActionResult> IsInputActive(int inputNumber)
        {
            // TODO: Implement input active check
            await Task.Delay(1);
            return Ok(new { inputNumber, isActive = false });
        }

        /// <summary>
        /// Check if specific output is active
        /// </summary>
        /// <param name="outputNumber">Output number to check</param>
        /// <returns>Output active status</returns>
        [HttpGet("IsOutputActive/{outputNumber}")]
        public async Task<IActionResult> IsOutputActive(int outputNumber)
        {
            // TODO: Implement output active check
            await Task.Delay(1);
            return Ok(new { outputNumber, isActive = false });
        }

        /// <summary>
        /// Set output state
        /// </summary>
        /// <param name="request">Output setting request</param>
        /// <returns>Success response</returns>
        [HttpPost("SetOutputState")]
        public async Task<IActionResult> SetOutputState([FromBody] SetOutputRequest request)
        {
            // TODO: Implement set output state functionality
            await Task.Delay(1);
            return Ok(new { message = $"Output {request.OutputNumber} set to {request.State}", outputNumber = request.OutputNumber, state = request.State });
        }

        #endregion

        #region Dust Boot Control

        /// <summary>
        /// Retract dust boot
        /// </summary>
        /// <returns>Success response</returns>
        [HttpPost("RetractDustBoot")]
        public async Task<IActionResult> RetractDustBoot()
        {
            // TODO: Implement retract dust boot functionality
            await Task.Delay(1);
            return Ok(new { message = "Dust boot retracted" });
        }

        /// <summary>
        /// Extend dust boot
        /// </summary>
        /// <returns>Success response</returns>
        [HttpPost("ExtendDustBoot")]
        public async Task<IActionResult> ExtendDustBoot()
        {
            // TODO: Implement extend dust boot functionality
            await Task.Delay(1);
            return Ok(new { message = "Dust boot extended" });
        }

        /// <summary>
        /// Check if dust boot is retracted
        /// </summary>
        /// <returns>Dust boot retracted status</returns>
        [HttpGet("IsDustBootRetracted")]
        public async Task<IActionResult> IsDustBootRetracted()
        {
            // TODO: Implement dust boot retracted check
            await Task.Delay(1);
            return Ok(new { isRetracted = true });
        }

        #endregion

        #region Vacuum Control

        /// <summary>
        /// Start vacuum
        /// </summary>
        /// <returns>Success response</returns>
        [HttpPost("StartVacuum")]
        public async Task<IActionResult> StartVacuum()
        {
            // TODO: Implement start vacuum functionality
            await Task.Delay(1);
            return Ok(new { message = "Vacuum started" });
        }

        /// <summary>
        /// Stop vacuum
        /// </summary>
        /// <returns>Success response</returns>
        [HttpPost("StopVacuum")]
        public async Task<IActionResult> StopVacuum()
        {
            // TODO: Implement stop vacuum functionality
            await Task.Delay(1);
            return Ok(new { message = "Vacuum stopped" });
        }

        /// <summary>
        /// Check if vacuum is on
        /// </summary>
        /// <returns>Vacuum on status</returns>
        [HttpGet("IsVacuumOn")]
        public async Task<IActionResult> IsVacuumOn()
        {
            // TODO: Implement vacuum on check
            await Task.Delay(1);
            return Ok(new { isOn = false });
        }

        #endregion

        #region Dust Collector Control

        /// <summary>
        /// Start dust collector
        /// </summary>
        /// <returns>Success response</returns>
        [HttpPost("StartDustCollector")]
        public async Task<IActionResult> StartDustCollector()
        {
            // TODO: Implement start dust collector functionality
            await Task.Delay(1);
            return Ok(new { message = "Dust collector started" });
        }

        /// <summary>
        /// Stop dust collector
        /// </summary>
        /// <returns>Success response</returns>
        [HttpPost("StopDustCollector")]
        public async Task<IActionResult> StopDustCollector()
        {
            // TODO: Implement stop dust collector functionality
            await Task.Delay(1);
            return Ok(new { message = "Dust collector stopped" });
        }

        /// <summary>
        /// Check if dust collector is on
        /// </summary>
        /// <returns>Dust collector on status</returns>
        [HttpGet("IsDustCollectorOn")]
        public async Task<IActionResult> IsDustCollectorOn()
        {
            // TODO: Implement dust collector on check
            await Task.Delay(1);
            return Ok(new { isOn = false });
        }

        #endregion

        #region Testing Methods (IO Overrides)

        /// <summary>
        /// Override input value (for testing purposes only)
        /// </summary>
        /// <param name="request">Input override request</param>
        /// <returns>Success response</returns>
        [HttpPost("OverrideInput")]
        public async Task<IActionResult> OverrideInput([FromBody] OverrideIORequest request)
        {
            // TODO: Implement input override functionality
            await Task.Delay(1);
            return Ok(new { message = $"Input {request.Number} overridden to {request.Value}", inputNumber = request.Number, value = request.Value });
        }

        /// <summary>
        /// Reset all input overrides
        /// </summary>
        /// <returns>Success response</returns>
        [HttpPost("ResetInputOverrides")]
        public async Task<IActionResult> ResetInputOverrides()
        {
            // TODO: Implement reset input overrides functionality
            await Task.Delay(1);
            return Ok(new { message = "Input overrides reset" });
        }

        /// <summary>
        /// Override output value (for testing purposes only)
        /// </summary>
        /// <param name="request">Output override request</param>
        /// <returns>Success response</returns>
        [HttpPost("OverrideOutput")]
        public async Task<IActionResult> OverrideOutput([FromBody] OverrideIORequest request)
        {
            // TODO: Implement output override functionality
            await Task.Delay(1);
            return Ok(new { message = $"Output {request.Number} overridden to {request.Value}", outputNumber = request.Number, value = request.Value });
        }

        /// <summary>
        /// Reset all output overrides
        /// </summary>
        /// <returns>Success response</returns>
        [HttpPost("ResetOutputOverrides")]
        public async Task<IActionResult> ResetOutputOverrides()
        {
            // TODO: Implement reset output overrides functionality
            await Task.Delay(1);
            return Ok(new { message = "Output overrides reset" });
        }

        #endregion
    }
}