# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "HavenCNC Server API",
    "description": "REST API for CNC server management hosted within WinForms application",
    "contact": {
      "name": "HavenCNC Server Team"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/CNCIO/IsInputActive/{inputNumber}": {
      "get": {
        "tags": [
          "CNCIO"
        ],
        "summary": "Check if specific input is active",
        "parameters": [
          {
            "name": "inputNumber",
            "in": "path",
            "description": "Input number to check",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/CNCIO/IsOutputActive/{outputNumber}": {
      "get": {
        "tags": [
          "CNCIO"
        ],
        "summary": "Check if specific output is active",
        "parameters": [
          {
            "name": "outputNumber",
            "in": "path",
            "description": "Output number to check",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/CNCIO/SetOutputState": {
      "post": {
        "tags": [
          "CNCIO"
        ],
        "summary": "Set output state",
        "requestBody": {
          "description": "Output setting request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetOutputRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SetOutputRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SetOutputRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/CNCIO/OverrideInput": {
      "post": {
        "tags": [
          "CNCIO"
        ],
        "summary": "Override input value (for testing purposes only)",
        "requestBody": {
          "description": "Input override request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OverrideIORequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/OverrideIORequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/OverrideIORequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/CNCIO/OverrideOutput": {
      "post": {
        "tags": [
          "CNCIO"
        ],
        "summary": "Override output value (for testing purposes only)",
        "requestBody": {
          "description": "Output override request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OverrideIORequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/OverrideIORequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/OverrideIORequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "OverrideIORequest": {
        "type": "object",
        "properties": {
          "number": {
            "type": "integer",
            "description": "Input/Output number",
            "format": "int32"
          },
          "value": {
            "type": "boolean",
            "description": "Override value"
          }
        },
        "additionalProperties": false,
        "description": "Request model for input/output overrides (testing only)"
      },
...
      "SetOutputRequest": {
        "type": "object",
        "properties": {
          "outputNumber": {
            "type": "integer",
            "description": "Output number",
            "format": "int32"
          },
          "state": {
            "type": "boolean",
            "description": "Desired state (true = on, false = off)"
          }
        },
        "additionalProperties": false,
        "description": "Request model for setting output state"
      }
    }
  }
}